pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["cherrydan <cherrydan.1978@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
